refactor(dbt): expose points_earned/numeric_grade_earned on fct_grades_assignments
Closes #3688 (mart portion). Drops polymorphic score/score_type in
favor of Ed-Fi-aligned typed columns. Cube schema update for the
removed columns is tracked separately (out of dbt scope).

Also fixes a SQL/YAML type drift in
int_powerschool__gradebook_assignments_scores: numeric_grade_earned
was declared float64 in YAML but cast to NUMERIC in SQL; aligning the
SQL cast to float64 unblocks the mart contract.

+++ src/dbt/kipptaf/models/powerschool/intermediate/int_powerschool__gradebook_assignments_scores.sql
@@ -54,7 +54,7 @@ with
 
             if(
                 a.scoretype in ('PERCENT', 'GRADESCALE', 'COLLECTED'),
-                safe_cast(s.actualscoreentered as numeric),
+                safe_cast(s.actualscoreentered as float64),
                 null
             ) as numeric_grade_earned,
